pearl-lolo-ai-agent/requirements.txt

## **FILE 2: requirements.txt**
```txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
langchain>=0.0.300
chromadb>=0.4.0
faiss-cpu>=1.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
pyyaml>=6.0
pillow>=10.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.0
pymupdf>=1.22.0
python-docx>=1.1.0
openpyxl>=3.1.0
ollama>=0.1.0
google-search>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0